refactor: mv download*.go to its own asset pkg
